requirements.txt
streamlit==1.35.0
pymongo==4.7.2
duckdb==0.10.2
python-dotenv==1.0.1
firebase-admin==6.5.0
langchain==0.2.0
langchain-openai==0.1.6
langchain-groq==0.1.4
pydantic==2.7.1
